requirements.txt
# ═══════════════════════════════════════════════
# SC2 Bot Core
# ═══════════════════════════════════════════════
burnysc2>=5.0.0
numpy>=1.20.0
matplotlib>=3.3.0
psutil>=5.8.0
s2clientprotocol>=4.19.0.0
sc2reader>=0.8.0
pyyaml>=6.0

# ═══════════════════════════════════════════════
# Crypto Trading Service
# ═══════════════════════════════════════════════
aiohttp>=3.9.0
pyupbit>=0.2.33
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=1.0.0
cryptography>=41.0.0

# ═══════════════════════════════════════════════
# Discord Bot (JARVIS)
# ═══════════════════════════════════════════════
discord.py>=2.3.0

# ═══════════════════════════════════════════════
# AI / GenAI
# ═══════════════════════════════════════════════
google-generativeai>=0.5.0

# ═══════════════════════════════════════════════
# MCP Servers
# ═══════════════════════════════════════════════
mcp>=1.0.0

# ═══════════════════════════════════════════════
# Web Scraping / Search
# ═══════════════════════════════════════════════
feedparser>=6.0.0
duckduckgo-search>=4.0.0

# ═══════════════════════════════════════════════
# Google APIs
# ═══════════════════════════════════════════════
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

# ═══════════════════════════════════════════════
# Text-to-Speech
# ═══════════════════════════════════════════════
edge-tts>=6.0.0
gTTS>=2.3.0

# ═══════════════════════════════════════════════
# System / Utilities
# ═══════════════════════════════════════════════
beautifulsoup4>=4.12.0
torch>=2.0.0
tqdm>=4.65.0
pyautogui>=0.9.54

# ═══════════════════════════════════════════════
# Media / Document Processing
# ═══════════════════════════════════════════════
Pillow>=10.0.0
PyPDF2>=3.0.0
yt-dlp>=2024.1.0

# ═══════════════════════════════════════════════
# Audio (Windows - volume control)
# ═══════════════════════════════════════════════
pycaw>=20230407; sys_platform == "win32"
comtypes>=1.2.0; sys_platform == "win32"
